BUILD: Instantiate thrust sort kernels only once
