+++ src/backend/cuda/kernel/thrust_sort_by_key.hpp
@@ -1,0 +1,13 @@
+#pragma once
+#include <Param.hpp>
+namespace cuda
+{
+    namespace kernel
+    {
+        ///////////////////////////////////////////////////////////////////////////
+        // Wrapper functions
+        ///////////////////////////////////////////////////////////////////////////
+        template<typename Tk, typename Tv>
+        void thrustSortByKey(Tk *keyPtr, Tv *valPtr, int elements, bool isAscending);
+    }
+}
